+++ src/InitialConditions.jl
@@ -189,7 +189,7 @@ end
 
 Divides the initial number density `num_Init` equally among momentum-space bins in the range of `pmin` to `pmax`, `umin` to `umax` and `hmin to hmax`. These ranges may be defined as either grid indices or physical values. This is then applies to the initial state vector `Initial`.
 """
-function InitialConstant!(Initial::Vector{F},PhaseSpace::PhaseSpaceStruct,species::String;pmin::T,pmax::T,umin::T=-1.0,umax::T=1.0,hmin::T=0.0,hmax::T=2.0,num_Init::AbstractFloat=1.0,x_idx::Int64=nothing,y_idx::Int64=nothing,z_idx::Int64=nothing,off_space_idx::Int64=nothing) where T <: Union{Float32,Float64,Int64} where F<:AbstractFloat
+function InitialConstant!(Initial::Vector{F},PhaseSpace::PhaseSpaceStruct,species::String;pmin::T,pmax::T,umin::T=-1.0,umax::T=1.0,hmin::T=0.0,hmax::T=2.0,num_Init::AbstractFloat=1.0,x_idx=nothing,y_idx=nothing,z_idx=nothing,off_space_idx=nothing) where T <: Union{Float32,Float64,Int64} where F<:AbstractFloat
 
     Momentum = PhaseSpace.Momentum
